fix: bring to parity with #6281
