--- src/tracing/config.test.ts
@@ -1,17 +1,0 @@
-import { ErrorEvent } from '@sentry/types'
-
-import { beforeSend } from './config'
-
-describe('beforeSend', () => {
-  const ERROR = {} as ErrorEvent
-  it('propagates an error', () => {
-    expect(beforeSend(ERROR, {})).toBe(ERROR)
-  })
-
-  it('omits eth_blockNumber failures', () => {
-    const originalException = new (class extends Error {
-      requestBody = JSON.stringify({ method: 'eth_blockNumber' })
-    })()
-    expect(beforeSend(ERROR, { originalException })).toBe(null)
-  })
-})
