+++ src/tracing/errors.ts
@@ -1,0 +1,23 @@
+import { ErrorEvent, EventHint } from '@sentry/types'
+
+/** Identifies ethers request errors (as thrown by {@type import(@ethersproject/web).fetchJson}). */
+function isEthersRequestError(error: Error): error is Error & { requestBody: string } {
+  return 'requestBody' in error && typeof (error as unknown as Record<'requestBody', unknown>).requestBody === 'string'
+}
+
+export function filterKnownErrors(event: ErrorEvent, hint: EventHint): ErrorEvent | null {
+  const error = hint.originalException
+  if (error instanceof Error) {
+    // ethers aggressively polls for block number, and it sometimes fails (whether spuriously or through rate-limiting).
+    // If block number polling, it should not be considered an exception.
+    if (isEthersRequestError(error)) {
+      const method = JSON.parse(error.requestBody).method
+      if (method === 'eth_blockNumber') return null
+    }
+
+    // If the error is a network change, it should not be considered an exception.
+    if (error.message.match(/underlying network changed/)) return null
+  }
+
+  return event
+}
